moved some functionality from mutual recursion analyzer
